requirements.txt
pyqt5~=5.15.2
pyqt5-tools
Pillow~=8.2.0